Fixs problem with NetworkType type after move to /common/types
