+++ src/module/settings/hook/useSelectedNetwork.ts
@@ -1,4 +1,5 @@
-import settingsState, { NetworkType } from "module/settings/state/SettingsState";
+import { NetworkType } from "module/common/types";
+import settingsState from "module/settings/state/SettingsState";
 import { useRecoilValue } from "recoil";
 
 export default function (): NetworkType {
