+++ src/module/wallet/component/input/NewAccountNameTextField/util/isAccountValid.ts
@@ -1,4 +1,4 @@
-import { NetworkType } from "module/settings/state/SettingsState";
+import { NetworkType } from "module/common/types";
 import { NearSDKService } from "near-peersyst-sdk";
 
 export function isAccountValid(name: string, network: NetworkType): boolean {
