+++ src/module/wallet/hook/useRecoverWallets.ts
@@ -1,7 +1,7 @@
-import { NetworkType } from "module/settings/state/SettingsState";
 import { useSetRecoilState } from "recoil";
 import walletState from "../state/WalletState";
 import WalletController from "../utils/WalletController";
+import { NetworkType } from "module/common/types";
 
 export default function useRecoverWallets() {
     const setWalletState = useSetRecoilState(walletState);
