+++ src/module/wallet/state/ServiceInstances/ServiceInstances.types.ts
@@ -1,4 +1,4 @@
-import { NetworkType } from "module/settings/state/SettingsState";
+import { NetworkType } from "module/common/types";
 import { NearSDKService } from "near-peersyst-sdk";
 
 export interface BaseNearSdkParams {
