+++ src/refactor/ui/config/validators/AddressValidator.ts
@@ -1,6 +1,6 @@
 import { BaseValidator } from "@peersyst/react-native-components";
 import { TranslateFn } from "@peersyst/react-native-components";
-import { NetworkType } from "module/settings/state/SettingsState";
+import { NetworkType } from "module/common/types";
 import { isAccountValid } from "module/wallet/component/input/NewAccountNameTextField/util/isAccountValid";
 
 export class AddressValidator extends BaseValidator {
